build.rs: load from url if local files not found

+++ build.rs
@@ -9,12 +9,21 @@ use reqwest::blocking::get;
 use tar::Archive;
 
 fn main() {
-    let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
     let arch = env::var("CARGO_CFG_TARGET_ARCH").unwrap();
+    let version = env::var("CARGO_PKG_VERSION").unwrap();
 
+    let dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR").unwrap());
+    let lib_dir = dir.join("staticlib").join("linux").join(arch.clone());
+    if lib_dir.exists() {
+        println!("cargo:rustc-link-search=native={}", lib_dir.display());
+        return
+    }
+
+
+    let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
     // Download the file
-    let mut response = get("https://github.com/btwiuse/pub/releases/download/v0.0.1/staticlib.tgz")
-        .expect("Failed to download file");
+    let url = format!("https://github.com/btwiuse/pub/releases/download/v{}/staticlib.tgz", version);
+    let mut response = get(url).expect("Failed to download file");
     let mut out = File::create(out_dir.join("staticlib.tgz")).expect("Failed to create file");
     copy(&mut response, &mut out).expect("Failed to copy content");
